requirements.txt
streamlit==1.39.0
pytesseract==0.3.13
Pillow==10.4.0
pandas==2.2.2
pdf2image==1.17.0

